refactor: rename tasks.router to notifications.router and update related references
