+++ src/api/tasks/notifications.router.js
@@ -16,33 +16,35 @@ const taskRouter = () => {
 
   //apply any middleware
   //router.use(viewDeleteRegistrationAuth);
-  // SEND EMAILS AND STUFF
+
+  // Send unsent emails for all registrations
   router.get("/bulk/sendnotification", async (req, res) => {
-    logEmitter.emit("functionCall", "tasks.router", "bulk/sendnotification");
+    logEmitter.emit("functionCall", "notifications.router", "bulk/sendnotification");
     let dryrun = !!req.query.dryrun;
     let throttle = req.query && req.query.throttle ? req.query.throttle : 500;
 
     try {
       await sendAllNotificationsForRegistrationsAction(req, res, dryrun, throttle);
     } catch (e) {
-      logEmitter.emit("functionFail", "tasks.router", "bulk/sendnotification");
+      logEmitter.emit("functionFail", "notifications.router", "bulk/sendnotification");
       await fail(406, res, e.message);
       throw e;
     }
-    logEmitter.emit("functionSuccess", "tasks.router", "bulk/sendnotification");
+    logEmitter.emit("functionSuccess", "notifications.router", "bulk/sendnotification");
   });
 
+  // Send unsent emails for a single registrations
   router.get("/sendnotification/:fsaId", async (req, res) => {
-    logEmitter.emit("functionCall", "tasks.router", "sendnotification/:fsaId");
+    logEmitter.emit("functionCall", "notifications.router", "sendnotification/:fsaId");
     const { fsaId = null } = req.params;
     try {
       await sendNotificationsForRegistrationAction(fsaId, req, res);
     } catch (e) {
-      logEmitter.emit("functionFail", "tasks.router", "sendnotification/:fsaId");
+      logEmitter.emit("functionFail", "notifications.router", "sendnotification/:fsaId");
       await fail(406, res, e.message);
       throw e;
     }
-    logEmitter.emit("functionSuccess", "tasks.router", "sendnotification/:fsaId");
+    logEmitter.emit("functionSuccess", "notifications.router", "sendnotification/:fsaId");
   });
 
   return router;
